requirements.txt
AppKit==0.2.8
attrs==22.2.0
beautifulsoup4==4.12.2
bokeh==3.3.1
boto3==1.29.5
botocore==1.32.5
bpython==0.24
brotli==1.1.0
brotlicffi==1.1.0.0
Carbon==1.1.10
cftime==1.6.3
ConfigParser==6.0.0
contextlib2==21.6.0
coreschema==0.0.4
cx_Oracle==8.3.0
Cython==3.0.5
dask==2023.11.0
defcon==0.10.3
dl==0.1.0
docutils==0.20.1
eventlet==0.33.3
fastparquet==2023.10.1
Foundation==0.1.0a0.dev1
freetype_py==2.4.0
fs==2.4.16
fsspec==2023.10.0
genapi==0.0.8
gevent==23.9.1
glyphsLib==6.5.0
greenlet==3.0.1
gssapi==1.8.3
HTMLParser==0.0.2
hypothesis==6.90.0
inotify==0.2.10
invoke==2.2.0
ipdb==0.13.13
ipython==8.17.2
ipywidgets==8.1.1
jnius==1.1.0
keyring==23.5.0
lockfile==0.12.2
lz4==4.3.2
matplotlib==3.7.1
mock==5.1.0
mtrand==0.1
munkres==1.1.4
mypy==1.7.0
MySQL-python==1.2.5
nose==1.3.7
numarray==1.5.1
Numeric==24.2
numexpr==2.8.7
odfpy==1.4.1
olefile==0.46
openpyxl==3.1.2
PasteScript==3.3.0
pickle5==0.0.12
pox==0.3.3
protobuf==3.12.4
psutil==5.9.0
pyasn1==0.5.1
PyInstaller==6.2.0
pylibmc==1.6.3
pymemcache==4.0.0
pymysql==1.1.0
pyOpenSSL==23.3.0
PyQt4==4.11.4
PyQt5==5.15.10
PyQt6==6.6.0
PySide2==5.15.2.1
PySide6==6.6.0
pytest==7.2.1
python_socks==2.4.3
pywatchman==1.4.1
pyxlsb==1.0.10
qtpy==2.4.1
Quartz==0.0.1.dev0
railroad==0.5.0
reportlab==3.6.8
Res==0.1.7
s3fs==2023.10.0
scikit_learn==1.3.0
scipy==1.11.4
setproctitle==1.3.3
sets==0.3.2
simplejson==3.19.2
Sphinx==7.2.6
SQLAlchemy==2.0.23
sspi==0.1.0
statsmodels==0.14.0
sympy==1.12
tblib==3.0.0
threadpoolctl==3.2.0
tornado==6.3.3
traitlets==5.13.0
trove_classifiers==2023.11.14
truststore==0.8.0
ufoLib2==0.16.0
uharfbuzz==0.37.3
unicodedata2==15.1.0
uritemplate==4.1.1
urllib3_secure_extra==0.1.0
websockets==12.0
win_inet_pton==1.1.0
wsaccel==0.6.6
xarray==2023.11.0
xattr==1.0.0
xlrd==2.0.1
xlsxwriter==3.1.9
xlwt==1.3.0
xmlrpclib==1.0.1
zopfli==0.2.3
django-recaptcha